requirements.txt
spotipy
python-dotenv
transformers
torch
torchaudio
protobuf==3.20.*
